+++ stable_baselines3/common/vec_env/vec_normalize.py
@@ -289,8 +289,3 @@ class VecNormalize(VecEnvWrapper):
         """
         with open(save_path, "wb") as file_handler:
             pickle.dump(self, file_handler)
-
-    @property
-    def ret(self) -> np.ndarray:
-        warnings.warn("`VecNormalize` `ret` attribute is deprecated. Please use `returns` instead.", DeprecationWarning)
-        return self.returns